requirements.txt
fastapi
uvicorn
groq
python-dotenv
transformers